+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### Bugfixes
 
+- [Issue #328](https://github.com/influxdb/influxdb/issues/328). Join queries with math expressions don't work
+
 ## v0.5.6 [2014-04-08]
 
 ### Features