+++ CHANGELOG.md
@@ -5,7 +5,8 @@
 - [Issue #328](https://github.com/influxdb/influxdb/issues/328). Join queries with math expressions don't work
 - [Issue #440](https://github.com/influxdb/influxdb/issues/440). Heartbeat timeouts in logs
 - [Issue #442](https://github.com/influxdb/influxdb/issues/442). shouldQuerySequentially didn't work as expected
-  causing count(*) queries on large time series to use lots of memory
+  causing count(*) queries on large time series to use
+  lots of memory
 
 ## v0.5.6 [2014-04-08]